refactor: 피드백 반영

+++ src/main/java/com/blackdog/springbootjpa/domain/order/repository/OrderRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.Optional;
 
-public interface OrderRepository extends JpaRepository<Order, Integer> {
+public interface OrderRepository extends JpaRepository<Order, Long> {
     @Query("SELECT o FROM Order o JOIN FETCH o.orderItems oi JOIN FETCH oi.item WHERE o.id = :orderId")
-    Optional<Order> findOrderWithDetails(@Param("orderId") int orderId);
+    Optional<Order> findOrderWithDetails(@Param("orderId") long orderId);
 }
